# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation",
        "title": "API Documentation",
        "contact": {
            "name": "Prosper Abouar",
            "email": "prosper.abouar@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/auth/activate": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Activate account"
                ],
                "summary": "Activate new user account",
                "parameters": [
                    {
                        "description": "Enter your information",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_request.ActivateAccountRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_response.ActivateAccountResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid token!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    },
                    "403": {
                        "description": "User account is already activated!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/reset/code": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reset password"
                ],
                "summary": "Reset password set code",
                "parameters": [
                    {
                        "description": "Enter your information",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_request.ResetPasswordCodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_response.ResetPasswordInitResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid token!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/reset/init-email": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reset password"
                ],
                "summary": "Reset password  with email init",
                "parameters": [
                    {
                        "description": "Enter your information",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_request.ResetPasswordEmailInitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_response.ResetPasswordInitResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid email input!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User with email not found!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/reset/init-phone": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reset password"
                ],
                "summary": "Reset password with phone number init",
                "parameters": [
                    {
                        "description": "Enter your information",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_request.ResetPasswordPhoneNumberInitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_response.ResetPasswordInitResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid phone number input!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User with phone number not found!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/reset/new-password": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Reset password"
                ],
                "summary": "Reset password set new password",
                "parameters": [
                    {
                        "description": "Enter your information",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_request.ResetPasswordNewPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_response.ResetPasswordNewPasswordResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid token or password input!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    },
                    "404": {
                        "description": "User not found!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/signin-email": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sign in"
                ],
                "summary": "Sign in user with email",
                "parameters": [
                    {
                        "description": "Enter your information",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_request.SignInEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_response.SignInResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid inputs!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Account not activated!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Invalid email or password!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/signin-phone": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sign in"
                ],
                "summary": "Sign in user with phone number",
                "parameters": [
                    {
                        "description": "Enter your information",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_request.SignInPhoneNumberRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_response.SignInResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid inputs!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Account not activated!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Invalid phone number or password!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/signin-provider": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sign in"
                ],
                "summary": "Sign in user with provider",
                "parameters": [
                    {
                        "description": "Enter your information",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_request.SignInWithProviderRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_response.SignInResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid inputs!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Invalid token!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/signup-email": {
            "post": {
                "security": [
                    {
                        "ApiKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sign up"
                ],
                "summary": "Sign up user with email",
                "parameters": [
                    {
                        "description": "Enter your information",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_request.SignUpEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_features_auth_data_response.SignUpResponse"
                        }
                    },
                    "302": {
                        "description": "User with this email already exists!",
                        "schema": {
                            "$ref": "#/definitions/github_com_4kpros_go-api_common_types.WebErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_4kpros_go-api_common_types.WebErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_4kpros_go-api_features_auth_data_request.ActivateAccountRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "github_com_4kpros_go-api_features_auth_data_request.ResetPasswordCodeRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "github_com_4kpros_go-api_features_auth_data_request.ResetPasswordEmailInitRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "github_com_4kpros_go-api_features_auth_data_request.ResetPasswordNewPasswordRequest": {
            "type": "object",
            "properties": {
                "newPassword": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "github_com_4kpros_go-api_features_auth_data_request.ResetPasswordPhoneNumberInitRequest": {
            "type": "object",
            "properties": {
                "phoneNumber": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_4kpros_go-api_features_auth_data_response.ActivateAccountResponse": {
            "type": "object",
            "properties": {
                "activatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_4kpros_go-api_features_auth_data_response.ResetPasswordInitResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "github_com_4kpros_go-api_features_auth_data_response.ResetPasswordNewPasswordResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Enter Bearer with space and your token",
            "type": "apiKey",
            "name": "Bearer",
            "in": "header"
        },
        "X-API-Key": {
            "description": "Enter the API key to have access",
            "type": "apiKey",
            "name": "ApiKey",
            "in": "header"
        }
    }
}